# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de carteiras digitais multi-tenant desenvolvido em Go",
        "title": "Suno Wallets API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ops/timeline": {
            "get": {
                "description": "Retorna timeline paginada das operações de um cliente, com filtros avançados por tickers, fontes, tipos de ativo e período. Suporta canonicalização de dados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations Timeline"
                ],
                "summary": "Timeline de operações do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "PETR4,VALE3",
                        "description": "Lista de tickers separados por vírgula",
                        "name": "tickers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "B3_RAW,USER_MANUAL",
                        "description": "Lista de fontes separadas por vírgula (B3_RAW, USER_MANUAL)",
                        "name": "sources",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "equity,fund",
                        "description": "Lista de tipos de ativo separados por vírgula",
                        "name": "assetTypes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01",
                        "description": "Data inicial no formato YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-12-31",
                        "description": "Data final no formato YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Aplicar canonicalização de dados (padrão: true)",
                        "name": "canonicalize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 100,
                        "description": "Tamanho da página (padrão: 100)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "abc123def456",
                        "description": "Cursor para paginação (keyset)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timeline paginada de operações",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "CPF obrigatório",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ops/timeline/export": {
            "get": {
                "description": "Exporta timeline completa de operações de um cliente em formato CSV ou JSON. Requer autenticação de admin e suporte a limite configurável.",
                "produces": [
                    "text/csv",
                    "application/json"
                ],
                "tags": [
                    "Operations Timeline"
                ],
                "summary": "Exportar timeline completa (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chave secreta de admin (se configurada)",
                        "name": "X-Admin-Secret",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "csv",
                        "description": "Formato do export (csv ou json, padrão: csv)",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 50000,
                        "description": "Limite de registros (padrão: 50000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo CSV ou dados JSON exportados",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "CPF obrigatório",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - requer permissão de admin",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "TenantID": {
            "description": "ID do inquilino para multi-tenancy",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        }
    }
}